refactor: упорядочил типы для работы с API-запросами
