+++ packages/client/src/api/request/utilits.ts
@@ -1,4 +1,4 @@
-import {TErrorAPI} from '../types';
+import {TErrorAPI} from '../typingAPI';
 
 export const isErrorAPI = (data: unknown | TErrorAPI): data is TErrorAPI => {
 	return (data as TErrorAPI).reason !== undefined;
